refactor utils folder
